# openapi/platform-api.json
{
  "components": {
    "schemas": {
...
      "BootstrapRequest": {
        "properties": {
          "adminDisplayName": {
            "type": "string"
          },
          "adminUsername": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        },
        "required": [
          "companyName",
          "adminUsername",
          "adminDisplayName",
          "password",
          "locale",
          "timezone"
        ],
        "type": "object"
      },
...
      "ChangePasswordRequest": {
        "properties": {
          "password": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "password"
        ],
        "type": "object"
      },
...
      "PublishWorkflowRequest": {
        "properties": {
          "environmentId": {
            "format": "uuid",
            "type": "string"
          },
          "workflowVersionId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "environmentId",
          "workflowVersionId"
        ],
        "type": "object"
      },
...
    },
...
  },
  "info": {
    "description": "",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Agentx Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/change-password": {
      "post": {
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/bootstrap": {
      "post": {
        "operationId": "bootstrap",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/workflows/{id}/deployments": {
      "get": {
        "operationId": "list_deployments",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "operationId": "publish",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "tags": [
          "workflows"
        ]
      }
    },
...
  },
...
}